test: separate non-finite PCB regression cases
