+++ tests/pcb/pcb-non-finite-text-and-keepout2.test.ts
@@ -13,34 +13,6 @@ const board: AnyCircuitElement = {
   thickness: 1.6,
 }
 
-test("non-finite silkscreen sizes use the default for normal and knockout text", () => {
-  const render = (font_size: number, is_knockout: boolean) =>
-    convertCircuitJsonToPcbSvg([
-      board,
-      {
-        type: "pcb_silkscreen_text",
-        pcb_silkscreen_text_id: "text_0",
-        pcb_component_id: "component_0",
-        anchor_position: { x: 1, y: 2 },
-        anchor_alignment: "top_left",
-        layer: "top",
-        font: "tscircuit2024",
-        text: "Default\nsize",
-        font_size,
-        is_knockout,
-      },
-    ]).replace(/silkscreen-knockout-mask-text_0-\d+/g, "knockout-mask")
-
-  for (const knockout of [false, true]) {
-    const expected = render(1, knockout)
-    expect(expected).not.toMatch(/NaN|Infinity/)
-    for (const invalid of [NaN, Infinity, -Infinity]) {
-      expect(render(invalid, knockout)).toBe(expected)
-    }
-    expect(render(2, knockout)).not.toBe(expected)
-  }
-})
-
 test("non-finite keepout radii render like zero without corrupting board bounds", () => {
   const render = (radius: number) =>
     convertCircuitJsonToPcbSvg([
